Finish exercises 4.21-4.22

+++ part4/blog/controllers/blog.js
@@ -10,11 +10,7 @@ blogsRouter.get('/', async (request, response) => {
 
 blogsRouter.post('/', async (request, response) => {
   const body = request.body
-  const decodedToken = jwt.verify(request.token, process.env.SECRET)
-  if (!decodedToken.id) {
-    return response.status(401).json({error: "token missing or invalid"})
-  }
-  const user = await User.findById(decodedToken.id)
+  const user = request.user
 
   if (body.title === undefined || body.author === undefined) {
     response.status(400).send({error: "blog must have title and author"})
@@ -36,8 +32,21 @@ blogsRouter.post('/', async (request, response) => {
 })
 
 blogsRouter.delete('/:id', async (request, response) => {
-  await Blog.findByIdAndRemove(request.params.id)
-  response.status(204).end()
+  const user = request.user
+  
+  // await Blog.findByIdAndRemove(request.params.id)
+  const blog = await Blog.findById(request.params.id)
+  if ( blog === undefined || blog === null ) {
+    response.status(404).send({error: `blog with id ${request.params.id} not found`})
+  }
+
+  if ( blog.user.toString() === user.id.toString() ) {
+    await Blog.findByIdAndRemove(request.params.id) 
+    response.status(204).end()
+  } else {
+    response.status(400).send({error: "invalid token"})
+  }
+
 })
 
 blogsRouter.put('/:id', async (request, response) => {
